+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+dev:
+  - fix issue where capella fork was not detected
+
 0.7.0:
   - speed up sync by only updating changed validators
   - add open telemetry tracing